refactor(team-builder): delete old identity-lane.tsx stub, update callers to import from ./identity
Final cleanup of the identity-lane refactor. Removes the
apps/web/src/components/team-builder/v2/lanes/identity-lane.tsx
re-export stub that was kept around during steps 6-8 to avoid breaking
imports.

Updated callers to import directly from ./lanes/identity (the barrel):
- poke-row.tsx
- lanes/active-row.tsx
- __tests__/identity-lane.test.tsx (real import)
- __tests__/poke-row.test.tsx (jest.mock target)
- __tests__/active-row.test.tsx (jest.mock target)

The identity-lane refactor is now complete: one dispatcher
(./identity/identity-lane.tsx), three layout components (SingleRow,
MidStack, Vertical) sharing the same atomic cells, and per-component
scoped CSS in identity-lane.module.css.
