+++ apps/web/src/components/team-builder/v2/__tests__/active-row.test.tsx
@@ -31,7 +31,7 @@ jest.mock("../lanes/calc-column", () => ({
   CalcColumn: () => <div data-testid="calc-column" />,
 }));
 
-jest.mock("../lanes/identity-lane", () => ({
+jest.mock("../lanes/identity", () => ({
   IdentityLane: ({
     pokemon,
     fieldErrors,
