+++ apps/web/src/components/team-builder/v2/__tests__/identity-lane.test.tsx
@@ -298,7 +298,7 @@ jest.mock("../format-gating", () => ({
 // Import after mocks
 // =============================================================================
 
-import { IdentityLane } from "../lanes/identity-lane";
+import { IdentityLane } from "../lanes/identity";
 import { type ValidationError } from "../../validation-hooks";
 import * as FormatGating from "../format-gating";
 import * as TrainersPokemon from "@trainers/pokemon";
