--- apps/web/src/components/team-builder/v2/lanes/identity-lane.tsx
@@ -1,6 +1,0 @@
-"use client";
-
-// Identity-lane has been split into per-shape layout components under
-// `./identity/`. This file is kept only to preserve the existing import
-// path until step 9 rewrites callers to import from `./identity` directly.
-export { IdentityLane } from "./identity";
